Revert change I made when I didn't know how /init worked
For reference, /init always parses init.rc when booting to any other mode
besides low power mode. It then reads the hardware name from /proc/cpuinfo
(the line starting with "Hardware") and converts it to lower case. In our
case, this is "smdkc110".
It then parses init.$HARDWARE.rc if it exists.